+++ src/main.js
@@ -7,7 +7,7 @@ import './assets/main.css'
 
 const app = createApp(App)
 
-app.config.globalProperties.app_version = 'v0.4.4';
+app.config.globalProperties.app_version = 'v0.5.0';
 
 app.use(router)
 app.use(InstantSearch)