+++ src/router/index.js
@@ -22,6 +22,12 @@ const router = createRouter({
             meta: { requiresAuth: false }
         },
         {
+            path: '/register',
+            name: 'register',
+            component: () => import('../views/RegisterView.vue'),
+            meta: { requiresAuth: false }
+        },
+        {
             path: '/',
             component: () => import('../AuthLayout.vue'),
             children: [
@@ -64,6 +70,11 @@ const router = createRouter({
                     path: '/shelf-cocktails',
                     name: 'shelf-cocktails',
                     component: () => import('../views/ShelfCocktailsView.vue'),
+                },
+                {
+                    path: '/profile',
+                    name: 'profile',
+                    component: () => import('../views/ProfileView.vue'),
                 }
             ]
         }