+++ src/views/ProfileView.vue
@@ -1,0 +1,9 @@
+<script setup>
+import Profile from '../components/Profile/Form.vue'
+</script>
+
+<template>
+  <main>
+    <Profile />
+  </main>
+</template>