+++ src/views/RegisterView.vue
@@ -1,0 +1,7 @@
+<script setup>
+import Register from '../components/Register.vue'
+</script>
+
+<template>
+  <Register />
+</template>